LangChain/requirements.txt
langchain==0.1.0
langchain-openai==0.0.2
python-dotenv==1.0.0
faiss-cpu==1.7.4
tiktoken==0.5.2 